+++ ObjCTests/Tests/LengthTests.m
@@ -5,7 +5,7 @@
 #import <Unitility/Length.h>
 #import <XCTest/XCTest.h>
 
-static CGFloat const kAccuracy = 0.01f;
+static float const kAccuracy = 0.01f;
 
 @interface LengthTests : XCTestCase
